Changed timestamp fields in the global trace information.
The starttime and endtime fields in the file containing global information
become difficult to swap if generated in a machine with different endianness.
Therefore, the nanosecond duration of the execution is calculated prior to dump
the file.

Note that nanoseconds are used although the time resolution is in microseconds.
This eases the calculus to the post-processing utility.